requirements.txt
Flask==2.3.2
gunicorn==20.1.0
face_recognition==1.3.0
numpy==1.24.3
Pillow==9.5.0